+++ code/xcvm/lib/core/src/defi/mod.rs
@@ -1,0 +1,2 @@
+//! so we need to swap/exchange on each chain with its own interface
+pub mod picasso;
