+++ code/xcvm/lib/core/src/substrate/mod.rs
@@ -1,3 +1,5 @@
+//! so we need just some code which helps to convert substrate and CW primitives back and forth
+
 /// Errors from integration of CosmwWasm <-> Substrate (types, conversions, encoding, host
 /// functions, etc)
 #[derive(thiserror::Error, Debug)]
